requirements.txt
python-telegram-bot
emoji~=1.4.2